separate remote controller
